Add files via upload
The Plant Pathology Challenge 2020 data set to classify foliar disease of apples
Citation: Thapa, R., K. Zhang, N. Snavely, S. Belongie, and A. Khan. 2020. The Plant Pathology Challenge 2020 data set to classify foliar disease of apples. Applications in Plant Sciences 8(9): e11390. doi:10.1002/aps3.1139
This study discusses the dataset that we are using and the competition that its from. 
It discusses the manner in which CNN can be used to classify the images. 
-	Using a standard CNN model, 97% accuracy could already be achieved. 
-	The exception is the “complex disease symp toms” category, which was classified with just 51% accuracy. The algorithm is available on GitHub (https://github.com/rthapa-26/ FGVC-Plant-Pathology-2020-challenge-dataset-)
It gives a good idea of what the different diseases look like. (with pictures) 
-	A majority of the pictures taken were of apple scab, cedar apple rust, Alternaria leaf spot, frogeye leaf spot, and healthy leaves 
It also discusses some of the challenges that we need to address such as: 
1.	(1) an imbalanced data set of different disease categories,
2.	(2) non-homogeneous image backgrounds, 
3.	(3) images taken at different times of day, 
4.	(4) images from plants at different maturity stages, 
5.	(5) images displaying multiple diseases, and 
6.	(6) images taken using different focus settings.